requirements.txt
pycrypto
pytesseract
requests